Initial checkout-refactoring commit
