+++ shared/Checkout/Routes/ThankYou/ThankYouPage.js
@@ -1,7 +1,7 @@
 import React from "react"
 import ReactDOM from "react-dom"
 
-const ThankYouPage = props => {
+const ThankYouPage = () => {
 	return (
 			<div className="thank-you">
 				<h1>Thank you!</h1>
